parse SEIS straight ray example into a SimPEG.SEIS.StraightRay Problem and Survey
