+++ SimPEG/SEIS/StraightRay/__init__.py
@@ -1,0 +1,2 @@
+from .StraightRayProblem import StraightRayProblem, lengthInCell
+from .StraightRaySurvey import StraightRaySurvey
