+++ docs/content/examples/SEIS_StraightRa.rst
@@ -1,0 +1,24 @@
+.. _examples_SEIS_StraightRa:
+
+.. --------------------------------- ..
+..                                   ..
+..    THIS FILE IS AUTO GENEREATED   ..
+..                                   ..
+..    SimPEG/Examples/__init__.py    ..
+..                                   ..
+.. --------------------------------- ..
+
+
+Seismic Straight Ray Tomography
+===============================
+
+
+
+.. plot::
+
+    from SimPEG import Examples
+    Examples.SEIS_StraightRa.run()
+
+.. literalinclude:: ../../../SimPEG/Examples/SEIS_StraightRa.py
+    :language: python
+    :linenos:
